+++ visu3d/plotly/fig_utils.py
@@ -172,9 +172,9 @@ def make_traces(
     if is_visualizable(val):
       if isinstance(val, dca.DataclassArray):
         if has_fig_config(val):
-          num_samples = val.fig_config.num_samples
-          if isinstance(num_samples, fig_config_utils.LazyValue):
-            num_samples = num_samples.value
+          num_samples = fig_config_utils.LazyValue.resolve(
+              val.fig_config.num_samples
+          )
           val = math.subsample(
               val,
               num_samples=num_samples,